# openapi.yaml
openapi: 3.0.0
info:
  title: RDV API
  description: Teacher Meeting Scheduler API with Authentication and Children Management
  version: 2.0.0
servers:
  - url: http://localhost:4000/api
    description: Development server - API endpoints
  - url: http://localhost:4000/auth
    description: Development server - Authentication endpoints
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Child:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        name:
          type: string
          description: Legacy name field (for backward compatibility)
        birthday:
          type: string
          format: date
          description: Child's birthday (YYYY-MM-DD format)
        parentId:
          type: string
        classId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /children:
    get:
      tags:
        - Children Management
      summary: List children (filtered by user role)
      parameters:
        - name: classId
          in: query
          schema:
            type: string
          description: Filter by class ID
      responses:
        '200':
          description: List of children
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Child'
    post:
      tags:
        - Children Management
      summary: Create a new child
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [classId]
              properties:
                firstName:
                  type: string
                  description: Child's first name (preferred)
                lastName:
                  type: string
                  description: Child's last name (preferred)
                name:
                  type: string
                  description: Full name (legacy format, will be split into firstName/lastName)
                birthday:
                  type: string
                  format: date
                  description: Child's birthday (YYYY-MM-DD format, optional)
                classId:
                  type: string
                parentId:
                  type: string
                  description: Required for administrators and class leads
      responses:
        '201':
          description: Child created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Child'
        '400':
          description: Invalid input
        '403':
          description: Insufficient permissions
  /children/{childId}:
    get:
      tags:
        - Children Management
      summary: Get a specific child
      parameters:
        - name: childId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Child details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Child'
        '404':
          description: Child not found
        '403':
          description: Access denied
    put:
      tags:
        - Children Management
      summary: Update a child
      parameters:
        - name: childId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                name:
                  type: string
                  description: Legacy name field
                birthday:
                  type: string
                  format: date
                  description: Child's birthday (YYYY-MM-DD format)
                classId:
                  type: string
      responses:
        '200':
          description: Child updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Child'
        '404':
          description: Child not found
        '403':
          description: Access denied
    delete:
      tags:
        - Children Management
      summary: Delete a child
      parameters:
        - name: childId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Child deleted successfully
        '404':
          description: Child not found
        '403':
          description: Access denied

  # API Endpoints
...